majorproject/infrastructure/monitoring/flowguard_exporter/requirements.txt
prometheus-client==0.22.1
psycopg2-binary==2.9.11
redis==5.3.1
